feature/update-mex-common-dependency-to-0.65 (#47)
### Changes
- update mex-common to 0.65

as the Backend API Connector now allows to use a function to fetch ALL
merged items, this is used where applicable. The according needed
surrounding code (like first finding out the total number of items to be
fetched), was therefore removed.

Also the tests needed to be adapted for using the new function and also
for changing the positional arguments to keyword arguments int connector
fetch functions.